Crate bones_camera_shake

This crate provides 2D camera shake using the methodology described in this excellent GDC talk by Squirrel Eiserloh.

Example

use bevy::prelude::*;
use bones_camera_shake::*;

fn main() {
    App::new()
        .add_plugins(DefaultPlugins)
        .add_plugin(CameraShakePlugin)
        .add_startup_system(spawn_camera);
}

fn spawn_camera(mut commands: Commands) {
    commands.spawn((
        Camera2dBundle::default(),
        CameraShake::new(90.0, Vec2::splat(100.0), 0.5),
    ));
}

fn add_camera_shake(mut ev_trauma: EventWriter<CameraTrauma>) {
    /* some traumatic event... */
    ev_trauma.send(CameraTrauma(0.5));
}

Structs

CameraShake

Component for an entity with camera shake.

CameraShakePlugin

Add this plugin to your app to enable camera shake.

CameraTrauma

Event to add trauma to the camera. Provide a value between 0 and 1 for the trauma amount.